Extract CSS helper to css.js
